requirements.txt
django==1.10.6
djangorestframework==3.6.2
django_rest_swagger==2.1.2
pika==0.10.0
schedule==0.4.3
requests==2.17.3